refactor: split prime module into smaller files
