+++ core/prime/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_SAFE_INTEGER_BIGINT = BigInt(Number.MAX_SAFE_INTEGER);
+export const MAX_SAFE_SQRT = BigInt(Math.floor(Math.sqrt(Number.MAX_SAFE_INTEGER)));
